LPS-159188 Copy over logic to their respective display contexts. Base class contains logic that is shared across all of the module
