+++ modules/apps/layout/layout-admin-web/src/main/java/com/liferay/layout/admin/web/internal/display/context/EditLayoutDisplayContext.java
@@ -1,0 +1,162 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.layout.admin.web.internal.display.context;
+
+import com.liferay.client.extension.type.manager.CETManager;
+import com.liferay.item.selector.ItemSelector;
+import com.liferay.layout.page.template.model.LayoutPageTemplateEntry;
+import com.liferay.layout.page.template.service.LayoutPageTemplateEntryLocalServiceUtil;
+import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.model.Layout;
+import com.liferay.portal.kernel.model.LayoutConstants;
+import com.liferay.portal.kernel.model.LayoutSet;
+import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
+import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
+import com.liferay.portal.kernel.util.HttpComponentsUtil;
+import com.liferay.staging.StagingGroupHelper;
+
+import java.util.Locale;
+import java.util.TreeMap;
+
+/**
+ * @author Jürgen Kappler
+ */
+public class EditLayoutDisplayContext extends BaseLayoutsAdminDisplayContext {
+
+	public EditLayoutDisplayContext(
+		CETManager cetManager, ItemSelector itemSelector,
+		LiferayPortletRequest liferayPortletRequest,
+		LiferayPortletResponse liferayPortletResponse,
+		StagingGroupHelper stagingGroupHelper) {
+
+		super(
+			cetManager, itemSelector, liferayPortletRequest,
+			liferayPortletResponse, stagingGroupHelper);
+	}
+
+	public String getConfigurationTitle(Layout layout, Locale locale) {
+		LayoutPageTemplateEntry layoutPageTemplateEntry =
+			LayoutPageTemplateEntryLocalServiceUtil.
+				fetchLayoutPageTemplateEntryByPlid(layout.getPlid());
+
+		if (layoutPageTemplateEntry != null) {
+			return layoutPageTemplateEntry.getName();
+		}
+
+		return layout.getName(locale);
+	}
+
+	public String getFriendlyURLBase() {
+		StringBuilder friendlyURLBase = new StringBuilder();
+
+		friendlyURLBase.append(themeDisplay.getPortalURL());
+
+		Layout selLayout = getSelLayout();
+
+		LayoutSet layoutSet = selLayout.getLayoutSet();
+
+		TreeMap<String, String> virtualHostnames =
+			layoutSet.getVirtualHostnames();
+
+		if (virtualHostnames.isEmpty() ||
+			!_matchesHostname(friendlyURLBase, virtualHostnames)) {
+
+			Group group = getGroup();
+
+			friendlyURLBase.append(
+				group.getPathFriendlyURL(isPrivateLayout(), themeDisplay));
+			friendlyURLBase.append(
+				HttpComponentsUtil.decodeURL(group.getFriendlyURL()));
+		}
+
+		return friendlyURLBase.toString();
+	}
+
+	public Long getLayoutId() {
+		if (_layoutId != null) {
+			return _layoutId;
+		}
+
+		_layoutId = LayoutConstants.DEFAULT_PARENT_LAYOUT_ID;
+
+		Layout selLayout = getSelLayout();
+
+		if (selLayout != null) {
+			_layoutId = selLayout.getLayoutId();
+		}
+
+		return _layoutId;
+	}
+
+	public String getRootNodeName() {
+		if (_rootNodeName != null) {
+			return _rootNodeName;
+		}
+
+		_rootNodeName = getRootNodeName(isPrivateLayout());
+
+		return _rootNodeName;
+	}
+
+	public String getRootNodeName(boolean privateLayout) {
+		Group liveGroup = getLiveGroup();
+
+		return liveGroup.getLayoutRootNodeName(
+			privateLayout, themeDisplay.getLocale());
+	}
+
+	public boolean isDraft() {
+		Layout layout = getSelLayout();
+
+		if (layout.isDraftLayout() && layout.isSystem()) {
+			return true;
+		}
+
+		return false;
+	}
+
+	public boolean isLayoutPageTemplateEntry() {
+		Layout layout = getSelLayout();
+
+		LayoutPageTemplateEntry layoutPageTemplateEntry =
+			LayoutPageTemplateEntryLocalServiceUtil.
+				fetchLayoutPageTemplateEntryByPlid(layout.getPlid());
+
+		if (layout.isTypeAssetDisplay() ||
+			((layoutPageTemplateEntry != null) && layout.isSystem())) {
+
+			return true;
+		}
+
+		return false;
+	}
+
+	private boolean _matchesHostname(
+		StringBuilder friendlyURLBase,
+		TreeMap<String, String> virtualHostnames) {
+
+		for (String virtualHostname : virtualHostnames.keySet()) {
+			if (friendlyURLBase.indexOf(virtualHostname) != -1) {
+				return true;
+			}
+		}
+
+		return false;
+	}
+
+	private Long _layoutId;
+	private String _rootNodeName;
+
+}
